refactor: remove unsed prefetching, make more readable edit content components
